Improve & fix tracing
